+++ lucabook/lib/luca_book.rb
@@ -1,11 +1,13 @@
 # frozen_string_literal: true
 
+require 'csv'
 require 'luca_record'
 require 'luca_book/version'
 
 module LucaBook
   autoload :Import, 'luca_book/import'
   autoload :Journal, 'luca_book/journal'
+  autoload :List, 'luca_book/list'
   autoload :Setup, 'luca_book/setup'
   autoload :State, 'luca_book/state'
 end